requirements.txt
django-split-settings==1.2.0
Django==3.2.21 # https://nvd.nist.gov/vuln/detail/CVE-2023-31047
djangorestframework==3.12.4
django-allauth==0.57.0
django-vite==2.1.3
django-structlog==4.0.0
python-dotenv==0.15.0
django-guardian==2.4.0
django-recaptcha==3.0.0
django-http-referrer-policy==1.1.1
django-rest-knox==4.0.1
black==22.6.0
django-cleanhtmlfield==1.1.0
django-debug-toolbar==3.2.3
django-permissions-policy==4.9.0
django-sendgrid-v5==1.2.2
flake8==6.0.0
gunicorn==20.0.4
isort==5.12.0
Pillow==9.1.0
pre-commit==2.20.0
pre-commit-hooks==4.4.0
psycopg2-binary==2.9.6
sendgrid==6.10.0
django-fsm==2.8.0
Pillow==9.1.0
graphviz==0.20.1
django-redis==5.4.0
django-querycount==0.8.3